Miner Account Recovery
Making miners recover accounts
notes:
  - needed to change Gridutils to make it compile, it expects Int64 instead of integer on my latest stable lazarus.

+++ src/gui-classic/UGridUtils.pas
@@ -298,7 +298,7 @@ Var
   LApplyfilter : Boolean;
   LAccount : TAccount;
   LNode : TNode;
-  LAccountsList : TList<Integer>;
+  LAccountsList : TList<Int64>;
 begin
   LApplyfilter := ((FAccountsGridFilter.MinBalance>0) Or ((FAccountsGridFilter.MaxBalance>=0) And (FAccountsGridFilter.MaxBalance<CT_MaxWalletAmount)));
   FBalance := 0;
@@ -321,7 +321,7 @@ begin
             LAccountsNumbersList := FAccountsGridFilter.OrderedAccountsKeyList.AccountKeyList[i];
             if Assigned(LAccountsNumbersList) then begin
 
-              LAccountsList := TList<Integer>.Create;
+              LAccountsList := TList<Int64>.Create;
               Try
                 LAccountsNumbersList.FillList(j_min,500,LAccountsList);
                 for j := 0 to LAccountsList.Count - 1 do begin
